add strict checking of endpoint URI
This commit changes the way endpoints are resolved.

If an endpoint isn't found, I think it resolves to "/" since that is
the matching prefix. This causes funky errors with bad payloads. As
such, this commit also does some stricter payload checking.

Lastly, the version information is needed all over the place, so we
load those earlier in a new file.