Feature: approximate compilation (#143)
The major feature of this commit is the introduction of "approximate compilation", replacing the previous contents of `src/compilers/optimal-2q.lisp`. This new suite of functions produces from all circuits of a particular shape the 'nearest', in the sense of average state fidelity, to a given program. This can then be used to either perform approximate compilation or to perform traditional optimal compilation (by rejecting all results with imperfect fidelity distance).

Some other minor features include:
* Improved numerical stability in `collinearp`. Hopefully this reduces test suite errors.
* The removal of various `environs` optional arguments, which have been essentially unused since the introduction of `named-operator` and ilk.
* The addition of a `rewiring` keyword argument to `make-matrix-from-quil`, to help cut down on mutation of `instruction` objects.