Reduce global state mutation in quilc
Replace most global state mutation in quilc with let bindings.

The only remaining global state mutation occurs in
quilc::setup-debugger which is called from quilc::entry-point, but not
quilc::%entry-point. Removing this last bit of state mutation is also
possible, but slightly uglier since it requires either peeking under
the covers of sb-ext:disable-debugger to figure out what additional
vars to let-bind, or else running the body of quilc::entry-point in an
unwind-protect that calls quilc::enable-debugger on exit. Since the
whole point of quilc::%entry-point is presumably to provide a
non-debugger-mangling entry point for interactive use, it didn't seem
worth it to take this final step.

This is "option 5" from the discussion in #192.

Fixes #192