efficient classical logic gate compilation with tweedledum (#150)
* mvp

This adds tweedledum as a git submodule to alleviate complexity in
locating it on a user's system, and makes the compilation process
painless.

See the comments in cl-tweedledum.lisp for an example of using it.

* restructure

* Refactor install-generic-compilers for gr8r cstmzblty

e.g. the tweedledum defgate compiler can be installed
via

(push (constantly #'cl-quil.tweedledum::compile-perm-gate-with-tweedledum) quil::*global-compilers*)

* holy smokes 3x

* remove debug print

* cl-tweedledum depends on cl-quil

* code smell comment

tech debt

* add DEFGATE NAME AS (MATRIX|PERMUTATION) to quil

Can elide the MATRIX type as default

* fix accidental backspace

* fix gitmodules path

* remove broken submodule

* trying to fix this submodule stuff

* push onto compilers when contrib is loaded

* attempt to handle missing c++17 comiler gracefully

* in case of error, do not offer to continue without contrib

If you try to load a module, and that module fails to load, then you should
probably take the hint to fix stuff rather than ignore it.

* more efficient derivation of permutation for a DEFGATE body

* fix permutation calculation

* cl-quil/tweedledum-tests

runs cl-quil-tests with tweedledum enabled

* attempt 4.1 at submodules

* submodules again

* update submodule

* fix test stuff

* add 4gotten file

* submodules are hard

* Add README.md

* more README.md

* Remove conditional in favour of note

* descriptive errro

* use merge-pathnames liberally

* review comments

* fix error

* plural were/was in error message

* use quil-parse-error rather than error

* fix some DEFGATE AS PERMUTATION parsing edge cases

* Some DEFGATE AS (MATRIX|PERMUTATION) tests

* Show gate type in error message, rather than token-type

i.e. when parsing "DEFGATE TEST AS SOMETHING: ...", the error should
be "Found unexpected gate type: SOMETHING." not "Found unexpected gate
type: NAME."

* remove contribs stuff

gonna have that in a separate PR

* load tweedledum into sdk image

* git submodule stuff for gitlab ci

* tweedledum updates

* output build errror

* .

* update tweedledum

now without c++17 dependency!

* attempt to install g++-7

* boop

* beep

* add clang-7 apt repo

* remove version dump info

* add tweedledum to mangle shared objects

* remove load from build-app.lisp

in the interest of merging this PR, we'll fix external build-related
stuff later

* parse-quil-string -> parse-quil

* Remove rebase remnant

* remove unused code

* revcom