portability: add ccl-specific fixes
CCL still can't load cl-quil because of adt:defdata using not interned
symbols. Problem was fixed in ECL with commit
https://gitlab.com/embeddable-common-lisp/ecl/merge_requests/152/diffs?commit_id=2750dc993b9e88500d96b288a7f34886759d5e02

To make it working on CCL all #.(named-operator …) entries should be
replaced with (named-operator …), but since I do not know how
essential it is I'm not going to change this part. Alternatively CCL
may learn how to make-load-form-saving-slots on structure-object so it
works on slot-offsets (instead of names).