Add hack to reset CL-QUIL::**COMMENTS** for testing purposes
Resetting the **COMMENTS** table might be required for any test that
wants to validate the output of PRINT-PARSED-PROGRAM for programs that
contain HALT, NOP, RESET, or other instructions that are represented
by singleton classes. This is because the **COMMENTS** hash table is
keyed on object identity, so if any prior tests have attached rewiring
comments to any singleton instructions, they will persist and be
printed by PRINT-PARSED-PROGRAM.

See also "Curio No. 2: Comment sharing" from
https://github.com/rigetti/quilc/issues/245