Add abstraction around quil<-->lisp function/operator conversion
The purpose of this commit is to add single source of truth + some
abstraction for how Quil functions and operators map to a lisp symbols
and vice versa.

The motivation for this change is that previously there were a handful
of places were the logic for validating and/or translating between
Quil and lisp was duplicated, which led to bugs when the list of valid
functions, say, was updated in one place but not the other.