Fix EXPT printing (#257)
* Fix printing of EXPT in delayed expressions

Ensure that the lisp function EXPT maps to the Quil infix operator "^"
when printing a DELAYED-EXPRESSION in PRINT-INSTRUCTION-GENERIC.

Fixes #249

* Add more PRINT-INSTRUCTION tests + support for testing golden files

- Move existing PRINT-INSTRUCTION tests into a new file
  tests/printer-tests.lisp

- Add support for testing "golden files", which are tests files that
  contain a sequence of input/expected-output pairs. See the comment
  in tests/utilities.lisp for more info.

- Add a bunch of new PRINT-INSTRUCTION tests using the new golden file
  format.

* Rename test-print-instruction to test-instruction-fmt

Rename the test and replace all

  (with-output-to-string (s)
    (cl-quil::print-instruction ... s))

with

  (format nil "~/cl-quil:instruction-fmt/" ...).

Updates #116

* Add hack to reset CL-QUIL::**COMMENTS** for testing purposes

Resetting the **COMMENTS** table might be required for any test that
wants to validate the output of PRINT-PARSED-PROGRAM for programs that
contain HALT, NOP, RESET, or other instructions that are represented
by singleton classes. This is because the **COMMENTS** hash table is
keyed on object identity, so if any prior tests have attached rewiring
comments to any singleton instructions, they will persist and be
printed by PRINT-PARSED-PROGRAM.

See also "Curio No. 2: Comment sharing" from
https://github.com/rigetti/quilc/issues/245

* Add abstraction around quil<-->lisp function/operator conversion

The purpose of this commit is to add single source of truth + some
abstraction for how Quil functions and operators map to a lisp symbols
and vice versa.

The motivation for this change is that previously there were a handful
of places were the logic for validating and/or translating between
Quil and lisp was duplicated, which led to bugs when the list of valid
functions, say, was updated in one place but not the other.

* Move generic deftypes and type predicates into a new types.lisp file

* Give credit where credit is due

Mention that printer-test-files/decomposed-ccnot.quil was stolen from
Nielsen & Chuang.

* Add mref test to printer-test-files/delayed-expressions.quil

* Add tests/printer-test-files/pragmas.quil

* Improvements to PARSE-AND-PRINT-QUIL-TO-STRING

Allow the caller of PARSE-AND-PRINT-QUIL-TO-STRING to specify the
:PARSER and :PRINTER they want, and make sure to use this function
everywhere applicable in printer-tests.lisp.

* More printer-tests tweaks

- Allow a golden test case to disable the fixed-point check for a
  particular output by leaving a comment in the corresponding Input
  section.

- Add a golden test file defgates.quil that contains an example of an
  input/output pair for which the fixed-point check needs to be
  disabled.

- Move one of the test cases that previously appeared in
  TEST-DEFGATE-PRINTING into delayed-expressions.quil.

- Add comments explaining why the remaining test cases in
  TEST-DEFGATE-PRINTING and TEST-CIRCUIT-AND-DECLARE-PRINTING are not
  simply included in the golden parse/print tests.

- Update TEST-DEFGATE-PRINTING with a NOT-SIGNALS fiasco assertion.

* Add a DEFGATE ... AS PERMUTATION test to the golden printer tests

* Add TEST-JUMP-TO-INTEGER-LABEL-PRINTING

* Track originating file and line number when parsing golden files

- Add GOLDEN-TEST-CASE structure to represent individual input/output
  pairs parsed from a golden test file. This structure also tracks the
  originating file name and line number for each test, as well as the
  input and output sections.

- PARSE-GOLDEN-FILE{,-STREAM} now returns a list of GOLDEN-TEST-CASE
  structures.

- TEST-PRINT-PARSED-PROGRAM-GOLDEN-FILES now includes the file name
  and line number of the input/output pair that caused a failure for
  any failing assertions.

* Add MAP-GOLDEN-FILES-AND-TEST-CASES

This is a convenience wrapper for writing golden file tests that
handles iterating over golden files and any GOLDEN-TEST-CASEs they
include, as well as reporting status about the current file and test
case being processed.

* Add descriptive names for DEFGATE/DEFCIRCUIT golden tests

This aids debuggability in case of failing tests.

* Prefix comments with ;;; not ;; inside eval-when

* Convert remaining alexandria: references to spiffy new nickname a: