Construct hashtable directly rather than NativeQuilMetadata
This allows us to extend the returned metadata without having to
change the definition of rpcq::|NativeQuilMetadata|. This does have a
visible (though I believe unimportant) effect in pyQuil-world: the
metadata object returned by the `quil_to_native_quil` procedure will
be a real hashtable, rather than a `NativeQuilMetadata` object. The
`NativeQuilMetadata` was a thin wrapper around a hashtable anyway, and
supported all typical methods. This change ought not be BC
incompatible but who knows.