Provide estimated runtime in compiled protoquil metadata (#295)
* Replace references to cl-quil.clifford::extract-qubits-used

For details: https://github.com/rigetti/quilc/pull/157

* rename test to include "endpoint"

* rem declare

* binary exe resp endpoint

* review comms

* +-endpoint to name

* test remaining endpoints

* reduce duplication + random rpc clients

* keep appleby happy

* o m g

* Construct hashtable directly rather than NativeQuilMetadata

This allows us to extend the returned metadata without having to
change the definition of rpcq::|NativeQuilMetadata|. This does have a
visible (though I believe unimportant) effect in pyQuil-world: the
metadata object returned by the `quil_to_native_quil` procedure will
be a real hashtable, rather than a `NativeQuilMetadata` object. The
`NativeQuilMetadata` was a thin wrapper around a hashtable anyway, and
supported all typical methods. This change ought not be BC
incompatible but who knows.

* Convert "program_duration" to float before serializing

RPCQ's serialization routines don't like rationals.

* Compute estimated runtime

Note this is separate to the "program_duration". It might replace
that, but is separate for now.

* actually compute the runtime

* Implement PROTOQUIL-PROGRAM-P

CHECK-PROTOQUIL-PROGRAM now raises condition NOT-PROTOQUIL that should
be caught wherever it can occur (currently only in test suite).

* loop-de-loop

* Go back to using NativeQuilMetadata

* Fix type checking in runtime-estimation

* ignore this

* blerg

* rem prints

* use case

* fix nargs

* cond in place of if

* Remove old TODO

* don't forget to write to a stream

* (parantheses)