Fix printing of parameters in MATRIX-GATE-DEFINITION entries (#316)
* Fix printing of parameters in MATRIX-GATE-DEFINITION entries

If the entry is a symbol, it must be parameter. Pass it along to the
PRINT-INSTRUCTION method specialized on DELAYED-EXPRESSION, like we
already do for LISTs that represent Quil functions.

Fixes #272

* Remove unnecessary let-binding from PARSE-ARITHMETIC-TOKENS

* Move CL-PPCRE dependency out of CL-QUIL and into CL-QUIL-TESTS

* Plunder golden tests from good-test-files and compiler-hook-test-files

Do a one-time import of golden test cases for
TEST-PRINT-PARSED-PROGRAM-GOLDEN-FILES from
tests/{good,compiler-hook}-test-files.

Some of these plundered tests require regex support in order to verify
the printed output. Those tests have been sequestered in their own
gold-regex subdirectory, and the all the existing tests plus new
plundered tests that do not require regex support have been moved in
to a gold-standard subdir. The reason for separating the tests like
this is to avoid hoop-jumping when automatically updating golden file
output sections in UPDATE-PRINT-PARSED-PROGRAM-GOLDEN-FILES, which now
only updates the gold-standard files since I don't see an easy way to
automate updating the output sections of golden files that require
arbitrary regexes.

Closes #273

* Reverse parsed modifiers before calling APPLY-MODIFIERS-TO-OPERATOR

Fixes #317