Add some sanity checks for PARSE-QUIL -> P-P-T-L-M
Ensure that the combination of PARSE-QUIL and
PARSED-PROGRAM-TO-LOGICAL-MATRIX are working as expected, especially
w.r.t. gate modifiers and qubit consumption order.

This is change is motivated by my recent faux-pas [1] in #316
(reverted in #319) where I reversed the order in which gate modifiers
were parsed into an OPERATOR-DESCRIPTION without also permuting the
order of qubit arguments. That change broke the parser, but did it in
such a way that that no existing tests failed as a result.

[1]: https://github.com/rigetti/quilc/pull/316#discussion_r300177587