Back to the Reversal Part II (#329)
* Add some sanity checks for PARSE-QUIL -> P-P-T-L-M

Ensure that the combination of PARSE-QUIL and
PARSED-PROGRAM-TO-LOGICAL-MATRIX are working as expected, especially
w.r.t. gate modifiers and qubit consumption order.

This is change is motivated by my recent faux-pas [1] in #316
(reverted in #319) where I reversed the order in which gate modifiers
were parsed into an OPERATOR-DESCRIPTION without also permuting the
order of qubit arguments. That change broke the parser, but did it in
such a way that that no existing tests failed as a result.

[1]: https://github.com/rigetti/quilc/pull/316#discussion_r300177587

* Test logical matrices more sanely

* Fix the tests

* Fix the fix

* Nix matrix-insanity golden file tests

* Tabulate matrix cols and switch PHASE param 0 -> pi

* Update file header comment

* Add more test cases

* Add logical-matrix tests for compiled programs, too

* Add some compiler-hook-test-files that include modified gates

* Rename test to reflect that it now also tests compiled programs

* Run the logical-matrix-sanity tests first

Tests like the compiler-hook-tests and others depend on the correct
functioning of PARSED-PROGRAM-TO-LOGICAL-MATRIX on uncompiled
programs. Run the logical-matrix-sanity tests first, to serve as a
canary in the coal mine in case something has gone horribly wrong.

* Update and expand comments in logical-matrix-sanity-tests