Refactor %STRIP-HALTS-RESPECTING-REWIRINGS
Take ecprigetti's sage advice from #375 and refactor
%STRIP-HALTS-RESPECTING-REWIRINGS to only deal with a single HALT
instruction at the end of the program, rather than stripping HALTs
anywhere in the program. For starters, as ecp mentions in the comment
thread of #375 [1], this function is only called in protoquil mode, so
there should only be one HALT at the end. Even if there were multiple
HALTs, stripping away a mid-program HALT is not likely to leave you
with an equivalent program, so it doesn't seem very useful to support
that, even in theory.

As noted in the original thread, this refactoring actually makes the
code slightly longer and doesn't really improve big-O complexity since
we still need to copy the vector to strip off the final HALT. However
it does simplify (subjective) the test case since it only needs to
test trailing HALTs. The test code is now slightly longer, but
exhaustively tests the "interesting" cases. The TEST-COMPATIBLE /
TEST-INCOMPATIBLE tests are also more declarative and therefore (imo)
easier to reason about what cases are actually being tested than the
previous incarnation.

[1]: https://github.com/rigetti/quilc/pull/375#discussion_r311787805