Use first link candidate as default, in case rewiring not filled.
The potential link candidates are places where inserting a swap will
lower the cost function. It may turn out that the involved qubits are
not filled-out in the current rewiring and so the optimization loop
would either error (building a gate with NIL as an argument) or not
select any of the candidates (the WHEN-LET would have a NIL for either
CONTROL or TARGET).

Since we always need a swap from the list of candidates, it makes
sense to default to the first link in the candidates list. The
remaining links will (a) be better and thus selected, (b) be worse and
thus skipped, or (c) have qubits that are unassigned in the rewiring
and thus should be skipped.

I think that's a reasonable compromise, but I've been wrong before.