Look over entire potential links list
Default link is the first candidate, best-cost-so-far is NIL.

During iteration, if possible the cost of the default link (first
candidate) will be computed and stored in best-cost-so-far. Otherwise
best-cost-so-far remains NIL.

The call to COST-< is guarded by a nil test on best-cost-so-far, and
so can never be called with a non "cost"-type object.