Use first link candidate as default, in case rewiring not filled. (#544)
* Use first link candidate as default, in case rewiring not filled.

The potential link candidates are places where inserting a swap will
lower the cost function. It may turn out that the involved qubits are
not filled-out in the current rewiring and so the optimization loop
would either error (building a gate with NIL as an argument) or not
select any of the candidates (the WHEN-LET would have a NIL for either
CONTROL or TARGET).

Since we always need a swap from the list of candidates, it makes
sense to default to the first link in the candidates list. The
remaining links will (a) be better and thus selected, (b) be worse and
thus skipped, or (c) have qubits that are unassigned in the rewiring
and thus should be skipped.

I think that's a reasonable compromise, but I've been wrong before.

* Look over entire potential links list

Default link is the first candidate, best-cost-so-far is NIL.

During iteration, if possible the cost of the default link (first
candidate) will be computed and stored in best-cost-so-far. Otherwise
best-cost-so-far remains NIL.

The call to COST-< is guarded by a nil test on best-cost-so-far, and
so can never be called with a non "cost"-type object.

* Add test case for the Aspen 7 swap bug