Be more tolerant of IMPOSSIBLY-FULL-RESOURCEs
Be more tolerant of IMPOSSIBLY-FULL-RESOURCEs in RESOURCE-DIFFERENCE
and RESOURCE-SUBSETP. Previously, if RESOURCE-DIFFERENCE was given an
IMPOSSIBLY-FULL-RESOURCE as it's first argument, it would signal
RESOURCE-COLLECTION-UNBOUNDED-NAMES. Now it only does so if the second
argument is not also "impossibly full".

RESOURCE-SUBSETP now explicitly checks for impossibly full resources,
and only delegates to RESOURCE-DIFFERENCE in cases where it is
guaranteed safe and returns NIL otherwise.

Closes #346