More accurately calculate subschedule
The "subschedule" we're calculating here is defined as, given some
schedule of instructions and a resource of interest, beginning from
the last instructions in the schedule find the largest chain of
instructions using *only* a particular resource.

For example, say our schedule on a linear chip is created from the
instructions

    H 0; H 1; CNOT 2 0; H 1; CNOT 0 1; X 0

whose dependecy graph looks something like

    H       H
    0       1
    |       |
  CNOT      |
    |\      H
    | \     1
    |  \    |
    |   \   |
    |    CNOT
    |    0  1
    |   /   |
    |  X    |
    |  |    |
    2  0    1

If the resource we are interested in is the qubit collection (0, 1),
then the subschedule is

    H 1; H 1; CNOT 0 1; X 0.

For the resource (0),

    X 0.

For the resource (1),

    NIL

because there is no subschedule (retreating from the last
instructions) that acts solely on qubit 1. Likewise for qubit 2.

For the resource (0, 2),

    X 0.

PRIOR TO THIS COMMIT, the subschedule would be calculated as

    H 0; CNOT 2 0; X 0

which is incorrect because the link between CNOT 2 0 and X 0 is
blocked by the illegal instruction CNOT 0 1.

fixy fixy

fixy2

comm

test

test

fix test

add test

fix timing issue