Add MAGICL constructor shims for default matrix type
This commit adds shims for all the MAGICL constructors to set the
default element type to '(COMPLEX DOUBLE-FLOAT). This is being done
remove the need to specify the type on every call inside CL-QUIL.
NOTE: To get a DOUBLE-FLOAT matrix you will now need to specify the TYPE.

The shims are a way of avoiding setting MAGICL::+DEFAULT-TENSOR-TYPE+
for the whole package, as that would then have global side effects,
affecting packages loaded afterwards (and likely cause errors).
A modification to MAGICL to allow for a default for each package to be
set was also considered but would have made calls too verbose.