allow INST to be used dynamically
This commit doesn't change the how any existing functionality works,
but does permit new functionality.

Currently, INST & co may only be used lexically within a WITH-INST,
and by extension, within DEFINE-COMPILER. If INST is used outside of
either of those contexts, there will be a _compile-time_ error it can
be detected, but definitely a run-time error.

This commit allows INST to be used anywhere within the dynamic extent
of WITH-INST (and DEFINE-COMPILER). This means it's now possible to
write functions like:

    (defun inst-hadamard (N)
      (dotimes (q N)
        (inst "H" () q)))

There are two downsides:

    - Compile-time errors cannot be detected since any INST call may
      possible be valid.

    - There may be a small performance penalty using dynamic
      features. However, what we lost in making it dynamic might be
      gained back by much smaller code. (INST and INST* aren't
      duplicated in every definition.)