Fix compiler path calculation
The computed compiler paths were:

  (1) not returning the shortest path, though the function claimed to do
  so; and
  (2) not including XY compilers, even if the target gateset included an
  XY gate that was better than a CZ gate on the same hardware object

This introduces a separate issue: arbitrarily long sequences of rx and
rz gates are not collapsed into at most five gates, but now collapse
into at most 7 gates. This is due to the compiler thinking that the
EULER-YZY compiler is better than the EULER-ZXZ compiler. I can live
with this regression for now, since the two extra gates don't cost much
in terms of fidelity, when compared to the improvement of allowing for
high-fidelity XY gates.

An extra nicety: we can now support SWAPs that compile to just two 2Q
gates (one XY and one CZ, cf. three XY or three CZ).