Account for number of gates of unknown fidelity in EVAL-OCC-TABLE
Previously, a gate with unknown fidelity and n occurrences was
recorded in the metric as having one unknown. This means an occurrence
table of {RX: 1, RY: 50} with RX known and RY unknown would have a
better metric than {RX: 2, RY: 1} since the RY in both cases is recorded
as 1 unknown. This causes issues in determining short compiler paths.

This change now records the former table with 50 unknowns
and the latter with 1 unknown. This change also adds additional
documentation to the OCCURRENCE-TABLE structure to make this clear.

TEST-RX-RZ-STRINGS-REDUCE has been returned to its initial expectation
of five gates.