Ansatz search (#670)
* add a new SOLVE-FOR-ANSATZ-INPUTS function

This commit adds a new function SOLVE-FOR-ANSATZ-INPUTS useful for
finding matrices in a given form. It just boils down to a Nelder-Mead
which is particularly effective in many cases. This function can be
used to build compilers for relatively exotic gate sets, provided
you've discovered a function which can represent them.

* fix spelling error

Co-authored-by: Erik Davis <erik@cadlag.org>

* simplify LET* to LET

Co-authored-by: Erik Davis <erik@cadlag.org>

Co-authored-by: Erik Davis <erik@cadlag.org>