Truncate fidelities to the range (0, 1)
i.e. the range [double-float-epsilon, 1 - double-float-epsilon]

We truncate perfect fidelities to less-than-perfect so that we have
some wiggle room when deciding what to do next. That less-than-perfect
fidelity was chosen to be 0.999, which I think (at least in theory)
could cause unexpected behaviour. Say for example you have one
instruction with a very good fidelity (say 0.999999) and another with
good fidelity (0.999). The compiler will truncate the former to 0.999,
and may then pick the second instruction for its output. The user is
then left thinking why is this instruction chosen over the superior
0.999999 instruction?

The user might then return to the compiler to inquire about the
resulting fidelity of the program, which it will report as 0.999 which
is fine if the compiler chose the less-good of the two options, but if
it chose the "better" of the two, then the user would expect to see a
fidelity of 0.999999.

This change goes some way to solving/alleviating those issues. In the
first case, 0.999999 will not be truncated and so the compiler will do
The Right Thing. In the second case, the fidelity will look reasonable
to the user. In the case where a fidelity of 1.0 is specified,
it *will* be truncated but it will still likely be larger than other
fidelities, and the fidelity calculation will still be reasonable to
the user (even if it isn't a perfect 1.0).