Fix test-memoize-pauli-basis-matrices timing issues.
Test was failing on a fast machine because the "slow" case took 1/2
msec or less on every iteration. Time-it macro would round the time
down to zero in an effort to return time in milliseconds, needlessly
throwing away the extra precision in internal time units.

There's no need for any particular time units (such as msec), since
the time is only used for comparison.

Furthermore, it should use get-internal-run-time, not -real-time,
i.e., should not be affected by other processes running.

Changed time-it to return time in terms of internal time units and to
be in terms of run time.  Also, added some documentation of what it
does.

Also, the test code only tested whether the slow case took longer at
all. Improved it to test that the slow test takes longer by some
minimum slow-to-fast time ratio (8x).

For comparison, ratios of 100x were typically achieved With SBCL,
standard optimization, on MacOS, on a Macbook Pro, 16-inch 2019, with
processor speed 2.6 GHz.

In the future, if machines get a lot faster, and/or a port is needed
to a fast Lisp with msec internal time units, basis-size could be
bumped higher, to slow the slow case.