Fix test TEST-BIG-DEFGATE stack overflow with certain compiler policy on SBCL.
With a certain ultra slow ultra safe compiler policy, a stack overflow
error during one of its tests, namely cl-quil-tests::test-big-defgate.

When compiler policy is set by evaluating the following

  (sb-ext:restrict-compiler-policy 'debug 3 3)
  (sb-ext:restrict-compiler-policy 'safety 3 3)
  (sb-ext:restrict-compiler-policy 'space 0 0)
  (sb-ext:restrict-compiler-policy 'speed 0 0)

the sytem gives the following error

    Control stack exhausted (no more space for function call frames).
    This is probably due to heavily nested or infinitely recursive function
    calls, or a tail call that SBCL cannot or has not optimized away.

There follows a backtrace with an extremely large number of calls to SPLIT-LINES.

This is evidently a result of tail call optimization not being done
due to having DEBUG so high and SPEED so low.

With this reimplementation of split-lines, i.e., without a recursive
inner function, the above test and all other tests pass without any
error no matter which compiler settings are used.