Add subsystem cl-quil/tools -- cl-quil language tools for developers
This is a stub for now, but we'll add to it later.

In anticipation of adding some dev-tools type modules, this introduces
a new subsystem cl-quil/tools that new modules can hang off of. For
now this is a stub, but later we may add, e.g., hasse-stuff (a little
utility for doing Hasse diagrams for logical schedulers) and circ-viz
(some tools for circuit visualization), maybe with different names.

The two anticipated new modules, and more to come later probably, are
not being added to cl-quil (core) system because they are
nice-to-haves but non-core. In addition, these two in particular have
weird dependencies like GraphViz and LaTeX, so cordoning them off
keeps the core a bit cleaner.

Issue "Print a logical-schedule as a Hasse diagram #712" relates
loosely to this in that Hasse stuff is planned to be added to this
stub system, so it was a motivater for this general approach.