[#723] fix error in test-measure-semantics due to change in qvm
This fixes issue 723 in quil (not qvm). A subfunction of test function
test-measure-semantics in quil (%make-density-qvm-initialized-in-basis
in tests/misc-tests.lisp) had been calling qvm system and qvm package
function make-density-qvm with :amplitudes keyword. It turns out that
this argument was being ignored. Its being ignored was apparently not
a problem for this test.  However, in a more recent version of QVM,
the keyword arg began being passed along to make-instance on
DENSITY-QVM, and this is not a supported init arg, so the test
produced an error message like the following:

  Invalid initialization argument:
    :AMPLITUDES
  in call for class #<STANDARD-CLASS QVM:DENSITY-QVM>.

The new code performs the desired functionality without the keyword
arg, and is actually simpler.

Resolves: #723

Note: this bug caused quilc CI failures due to the CI using qvm
version 1.17.2.  The CI was subsequently patched to use the older
version of qvm. This commit will work with both the old version
(v1.17.1) and the new (v1.17.2).

Note: the change in qvm that broke the quilc test was not caught by
the CI until now because quilc's CI used an older version of qvm
(qvm-v1.17.1) until just recently.  The version chosen by the CI is
determined by Quicklisp, which is loosely tied to version of QVM, but
my have a lag of weeks or longer.