fix COMPILATION-TOLERANCE slot missing error, resolves: #736
Error processing code accessed slots of the condition using WITH-SLOTS
macro and erroneously supplied slot names in the wrong package --
CL-QUIL-TESTS, not QUIL.  Rewrote to access slots using full accessor
forms, properly package qualified.  Note that the compiler does not
normally complain when WITH-SLOTS uses slot names that do not pertain
to the given class. It should be a bit safer now that we're using the
normal full accessors, since if these slot accessors were ever to
become invalid, it would be flagged at compile time.

Also added a new test for this machinery:
test-%with-loose-state-prep-compression-736