Generalize chips specs for higher order objects more thoroughly.
There were a couple places in the data structure definitions that
hardcoded assumptions about the order of the objects (to <1):

* The length of the OBJECTS vector in a CHIP-SPEC.
* The length of the CXNS vector in a HARDWARE-OBJECT.
* Various constructors and initializers for building complete chip
specifications.
* The computation of the lookup cache.

To solve this problem we introduce more general abstractions that
cover the use case of building a chip-specification, as well as making
on-demand adjustable vectors for those whose lengths depend on the
order of the objects.

To make sure this works we have an example chip spec consisting of
just 3 qubits and a 3 qubit gate on them, that can do CCNOT.