Be more precise about when to approximate single operation costs.
* Distinguish between the cases of a hardware object being
"dead" (which should strictly be declared in the chip specification),
and the case of an object simply not having a "compiler path". (i.e. a
sequence of compilation methods which sufficiently nativize given the
gate set on the object).
* Don't compute cost bounds for an instruction if doing so yields an
infinite cycle in the compiler, since it may be the case that the
compiler actually can nativize an instruction (e.g., but inserting the
appropriate swaps), but cannot do so during chip specification
warming.