remove unused keyword argument in defualt `do-greedy-addressing` method
The keyword argument `:initial-rewiring` appeared in the arglist for the default method,
and in the generic function's docstring, but it did not appear in the generic function's
declared arglist (nor did `&allow-other-keys`), and the default method never read from
`initial-rewiring`. References in the method arglist and the docstring are removed.