`locally notinline` declarations to prevent SBCL warnings
SBCL will warn when it compiles a call to a function for which an `inline` declaration is
in effect, but for which inlining is impossible. It will also warn when it compiles the
definition of a function for which an `inline` declaration is effect, but to which
not-inlined calls have previously been compiled.

Both of these cases appeared in define-compiler.lisp. CL-HEAP contains `inline`
declaimations for `enqueue` and `dequeue`, but those are generic functions and therefore
not elligible for inlining. Structure accessors are always `inline`, but define-compiler
forward-references accessors to the `gate-record` struct. Rather than addressing the root
of either of those problems, this commit adds local `notinline` declarations so SBCL will
not warn.