Improve optimization for the containers declared in types.lisp
No warning to fix here, but an opportunity for some minor missed compiler
optimizations. CL compilers are generally unable to do any useful optimization or
compile-time type checking with `satisfies` types, because they are incapable of
understanding the behavior of the `satisfies` function. The traditional solution is, for a
type FOO and a predicate function FOO-WITH-SOME-PROPERTY-P, to define the type
FOO-WITH-SOME-PROPERTY as `(and FOO (satisfies FOO-WITH-SOME-PROPERTY-P))`, not just
`(satisfies FOO-WITH-SOME-PROPERTY-P)`. That way, the compiler knows at compile time that
every `FOO-WITH-SOME-PROPERTY` is at the very least a `FOO`, and also that any `(not FOO)`
is also not a `FOO-WITH-SOME-PROPERTY`. Footgun: `FOO-WITH-SOME-PROPERTY-P` still needs to
check `(typep thing 'FOO)`.