Correct `copy-instance` nonconformance re: `defstruct`
The "default" method for `copy-instance` is now specialized on `standard-object` rather
than `t`. Aside from opening the possibility that it might have been applied to instances
of `built-in-class`, the default method's applying to `structure-object`s was
non-conformant and SBCL specific (possibly other implementations as well, but certainly
not universal), as the MOP as specified in AMOP applies only to `defclass`-defined
classes, not to `defstruct` classes.

In lieu of a generic implementation for structs, we now use a macro
`define-copy-struct-instance` to define methods specific to each `structure-object`
subclass.

Pros:
- More correct; no longer relying on SBCL-specific behavior
- Potential modest performance increases, as copying structure objects no longer uses the
  MOP to dynamically compute a set of slots to copy.

Cons:
- Ugly, complex macro-ology in `define-copy-struct-instance`
- Having to define instances explicitly for structs is frustrating
- Potential bugs when a `defstruct` is incompatibly redefined without also editing its
  corresponding `define-copy-struct-object` form.