Be smarter about which existing queue to flush.
In last commit, flushed all existing queues. I misread the original implementation, which in fact
only flushed the sub-queue which was in "passing" mode, which is the queue involving the greatest
number of qubits.

As of this commit, the output of the new COMPRESS-INSTRUCTIONS seems just as good as the original
implementation.