check for diagonal matrices in find-diag-in-e-basis
It seems that LAPACK becomes unstable on ARM when attempting to
find the eigenvectors and eigenvalues of a diagonal matrix. This
instability is triggered in find-diagonalizer-in-e-basis. The
operation becomes trivial when run on a diagonal matrix, so we
check for this and return the identity to avoid making a
potentially unstable LAPACK call.

Fixes #842.