Add: showing my work in sqisw decomp
NOTE: I will squash the commits on this branch and do a new PR to
eliminate commit noise.

In this commit, I have added several functions and tests in comment
blocks. These will be removed from the final PR. They are here to aid
reviewers by adding information about which questions have been so far
addressed, which functions can be counted on, and so on.

cheers