Simple repalcement: s/scheduler/schedule [tech-debt]
Rationale: there is an inconsistent usage between lscheduler and
lschedule, even in the naming of accessor functions. This commit
brings a consistent langauge to the source by opting for the terms
logical-schedule and lschedule instead of logical-scheduler and
lscheduler.

No external symbols have been affected. Tests for qvm, boondoggle,
cl-quil, and quilc all passed locally.