Cleanups having to do with fidelity of sets of instructions
Move defs near to use sites; refactor instrs fidelity

Fidelity calculations used by the compressor and by the fidelity
addresser were needlessly constructing logical schedules.

Peeking at the logical schedule fidelity calculations, they too have
been cleaned up and refatored for clarity and performance.

Generally opting for treating fidelity of a circuit as the minimum
fidelity of its constituent gates. Prior to this commit, when a composite
fidelity needed to fall in the range (0, 1] the calculation

(exp (- (sqrt (reduce #'+ fidelities
                 :key (lambda (f) (* (log f) (log f)))))))

was being used. For fidelities f in (0, 1] this produced a
non-negative value f0 that was at most the minimal f.